backend/requirements.txt
pandas
pandas-ta
yfinance
PyYAML
python-dotenv
matplotlib
requests
python-multipart
fastapi
uvicorn
pydantic

llama-index
llama-index-embeddings-openai
llama-index-llms-openai
openai

PyPDF2
arxiv
markdown

crewai
crewai-tools

starlette

python-jose
aiofiles